hw_27/customer_service/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
pika==1.3.2
python-dotenv==1.1.0